[Opinel] Factored some aws utils into aws
